AccountManagerPlugin: Doing unfinished business from my earlier coding days.
git-svn-id: https://trac-hacks.org/svn@10312 7322e99d-02ea-0310-aa39-e9a107903beb

+++ accountmanagerplugin/trunk/acct_mgr/admin.py
@@ -228,12 +228,9 @@ class AccountManagerAdminPages(Component):
                 else:
                     data['registration_error'] = _(
                         "The password store does not support creating users.")
-            elif req.args.get('remove'):
-                sel = req.args.get('sel')
-                if sel is None:
-                    # so nothing to be done
-                    pass
-                elif delete_enabled:
+            elif req.args.get('remove') and req.args.get('sel'):
+                if delete_enabled:
+                    sel = req.args.get('sel')
                     sel = isinstance(sel, list) and sel or [sel]
                     for account in sel:
                         acctmgr.delete_user(account)
